Fix initial post being deleted on Deadline (#6417)
## Motivation

In rare cases where building the initial ATX times out the initial PoST stored in the local DB is deleted. This leads to an endless loop of the node not being able to publish its first ATX because it expects an initial PoST but never generates one.

This PR fixes this issue.