Improve performance of proposal builder by not DB inserting active set if not necessary (#6422)
## Motivation

Closes #6418

The first proposal created by a node for any identity triggers persisting the activeset that is used by all identities of the node. This can be a slow process (especially when many ATXs are received by the node) and is usually not necessary, because the activeset is most likely already in the DB.

This PR moves persisting the active set to the beginning of an epoch and avoids a DB write if not necessary (which will usually be the case).