Use go version specified by go.mod for CI (#6560)
## Motivation

The CI still uses a hard coded go version to build and unit test the node code. Both the release and systest jobs already parse and use the version specified in `go.mod`. This updates the CI to also use the version in `go.mod` instead of a hard coded setting.